requirements.txt
Pillow-8.2.0
docx-0.2.4
lxml-4.6.3
python-docx-0.8.11